@@ -1,13 +1,13 @@
-const chai = require('chai');
-const rewire = require('rewire');
-const sinon = require('sinon');
-const { beforeEach } = require('mocha'); /* eslint-disable-line no-redeclare */
-const { EOL } = require('os');
+import chai, { expect } from 'chai'
+import sinonChai from 'sinon-chai'
+import sinon from 'sinon'
+import { beforeEach } from 'mocha'
+import { EOL } from 'node:os'
 
-chai.use(require('sinon-chai'));
+chai.use(sinonChai)
 
-const expect = chai.expect;
-const sut = rewire('../../src/handlers/pr-conventional-title');
+import sut from '../../src/handlers/pr-conventional-title.js'
+
 
 suite('Testing the pr-conventional-title handler', () => {
     suite('Test handler matching', () => {
@@ -27,13 +27,11 @@ suite('Testing the pr-conventional-title handler', () => {
     });
 
     suite('Test handler running', () => {
-        let loadSpy;
         let createCheckStub;
         let repoFuncStub;
         let updateCheckStub;
 
         let fakeContext;
-        let baseConfig;
 
         const fakeSha = '#f54dda543@';
         const fakeCheckId = 13;
@@ -73,12 +71,6 @@ suite('Testing the pr-conventional-title handler', () => {
             updateCheckStub.resolves();
             repoFuncStub = sinon.stub(); // stub for context.repo function to short-circuit return the expected response
             repoFuncStub.callsFake((a) => {return { ...getRepositoryInfoResponse, ...a }});
-            // wrap spy on load configuration
-            let loadConfig = sut.__get__('load');
-            loadSpy = sinon.spy(loadConfig);
-            sut.__set__('load', loadSpy);
-            // grab the default configuration for testing usage
-            baseConfig = sut.__get__('DEFAULT_CONFIG');
             // create a fake context for invoking the application with)
             fakeContext = Object.freeze({
                 octokit: {
@@ -109,8 +101,6 @@ suite('Testing the pr-conventional-title handler', () => {
             // then verify a check run was created and updated as expected
             expect(createCheckStub).to.have.been.calledOnceWith(expectedCreateCheckRunInfo);
             expect(updateCheckStub).to.have.been.calledOnceWith({...baseExpectedUpdateCheck, ...expectedOutput});
-            // verify custom configuration, if exists, is combined with base configuration
-            expect(loadSpy).to.have.been.calledOnceWith(optionalConfig ? {...baseConfig, ...optionalConfig} : baseConfig);
         }
 
         test('Test with a conventional pr title, expect a successful report', async () => {
